+++ lib/screens/profile_screen.dart
@@ -1,0 +1,46 @@
+import 'package:flutter/material.dart';
+
+import '../models/user.dart';
+
+class ProfileScreen extends StatelessWidget {
+  const ProfileScreen({
+    super.key,
+    required this.user,
+  });
+
+  final User user;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+        image: DecorationImage(
+          // user에서 데이터 가져오기
+          image: NetworkImage(user.backgroundImage),
+          fit: BoxFit.fitHeight,
+        ),
+      ),
+      child: Scaffold(
+        // 배경 투명
+        backgroundColor: Colors.transparent,
+        body: Column(
+          children: [
+            // 수직방향 Column 만들건데 Spacer()가 위의 남은 공간 차지
+            const Spacer(),
+            // 지름 110 짜리 프로필 이미지 구현
+            Container(
+              width: 110,
+              height: 110,
+              decoration: BoxDecoration(
+                shape: BoxShape.circle,
+                image: DecorationImage(
+                    image: NetworkImage(user.backgroundImage),
+                    fit: BoxFit.cover),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
